+++ fuel-merkle/src/common/node.rs
@@ -1,3 +1,4 @@
+use crate::common::{Bytes32, Bytes8};
 use core::{fmt, mem};
 
 pub trait Node {
@@ -28,7 +29,19 @@ where
 #[allow(type_alias_bounds)]
 pub type ChildResult<T: ParentNode> = Result<T, ChildError<T::Key, T::Error>>;
 
-pub struct ChildErrorKey<Key>(pub Key);
+pub struct ChildErrorKey<Key>(Key);
+
+impl fmt::Display for ChildErrorKey<&Bytes8> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", u64::from_be_bytes(*self.0))
+    }
+}
+
+impl fmt::Display for ChildErrorKey<&Bytes32> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", hex::encode(self.0))
+    }
+}
 
 #[derive(Debug, Clone)]
 #[cfg_attr(feature = "std", derive(thiserror::Error))]
